pyproject.toml
[tool.poetry]
name = "projeto-ed-satc"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
pandas = "^2.3.3"
psycopg2-binary = "^2.9.11"
faker = "^38.0.0"
numpy = "^2.0.0"
openpyxl = "^3.1.0"
python-dotenv = "^1.0.0"
tqdm = "^4.65.0"
requests = "^2.31.0"
python-dateutil = "^2.8.0"
pyyaml = "^6.0.0"


[tool.poetry.group.dev.dependencies]
isort = "^5.13.2"
blue = "^0.9.1"


[tool.poetry.group.docs.dependencies]
mkdocs = "^1.6.0"
mkdocs-material = "^9.5.23"
pymdown-extensions = "^10.8.1"
mkdocstrings = "^0.25.1"


[tool.poetry.group.test.dependencies]
pytest = "^8.2.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "faker>=38.0.0",
    "pandas>=2.3.3",
    "psycopg2-binary>=2.9.11",
    "numpy>=2.0.0",
    "openpyxl>=3.1.0",
    "python-dotenv>=1.0.0",
    "tqdm>=4.65.0",
    "requests>=2.31.0",
]
